src/oracle-tec/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "oracle-tec"
version = "0.1.0"
description = "A modern, OOP-based Oracle database client for the mingo-system, improving upon legacy cx_Oracle implementations."
dependencies = [
    "oracledb>=2.0.0",
    "python-dotenv>=1.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]
